refactor(log): common log namespace
