+++ src/DependencyInjection/EMSCommonExtension.php
@@ -22,6 +22,7 @@ class EMSCommonExtension extends Extension
     {
         $loader = new XmlFileLoader($container, new FileLocator(__DIR__.'/../Resources/config'));
         $loader->load('contracts.xml');
+        $loader->load('log.xml');
         $loader->load('services.xml');
         $loader->load('commands.xml');
         $loader->load('twig.xml');
