+++ src/Resources/config/services.xml
@@ -62,19 +62,6 @@
             <argument type="service" id="logger" />
             <argument type="service" id="ems_common.elastica.client" />
         </service>
-        <service id="ems_common.monolog.doctrine" class="EMS\CommonBundle\Helper\Logger\DoctrineHandler">
-            <argument type="service" id="doctrine.orm.default_entity_manager" />
-            <argument type="service" id="security.token_storage" />
-            <argument type="string">%ems_common.log_level%</argument>
-        </service>
-        <service id="ems_common.repository.log" class="EMS\CommonBundle\Repository\LogRepository" public="true">
-            <argument type="service" id="doctrine"/>
-        </service>
-        <service id="ems.logger" alias="ems_common.localized_logger"/>
-        <service id="ems_common.localized_logger" class="EMS\CommonBundle\Helper\Log\LocalizedLogger">
-            <argument type="service" id="logger" />
-            <argument type="service" id="translator" />
-        </service>
         <service id="ems.helper.admin_api" class="EMS\CommonBundle\Common\Admin\AdminHelper">
             <argument type="service" id="ems_common.core_api.factory" />
             <argument type="service" id="Psr\Cache\CacheItemPoolInterface" />
